pop_manage/requirements.txt
flask
deta
python-dotenv
werkzeug
pytz
flask-cors
requests